go through %zu for size_t's for those that go through our libbu functions (given we shouldn't rely on stdio func supporting %z).  better for avoiding type cast.  related to staching/unstashing pointers, go through %p
svn:revision:38799
svn:branch:trunk
svn:account:brlcad

+++ src/librt/primitives/dsp/dsp.c
@@ -4068,7 +4068,7 @@ get_obj_data(struct rt_dsp_internal *dsp_ip, const struct db_i *dbip)
 			     bip->u.uint16, in_cookie, bip->count);
 
 	if (got != bip->count) {
-	    bu_log("got %d != count %llu", got, (unsigned long long)bip->count);
+	    bu_log("got %d != count %zu", got, bip->count);
 	    bu_bomb("\n");
 	}
     }
